Revert "Allow permanent dataset layer rotation in dataset settings (#8159)"
This reverts commit 1fed6d30903dc9e52b1432d297f3b5ab42848ba3.
