+++ frontend/javascripts/admin/dataset/composition_wizard/04_configure_new_dataset.tsx
@@ -28,8 +28,7 @@ import _ from "lodash";
 import messages from "messages";
 import { WkDevFlags } from "oxalis/api/wk_dev";
 import type { Vector3 } from "oxalis/constants";
-import { getReadableURLPart } from "oxalis/model/accessors/dataset_accessor";
-import { flatToNestedMatrix } from "oxalis/model/accessors/dataset_layer_transformation_accessor";
+import { flatToNestedMatrix, getReadableURLPart } from "oxalis/model/accessors/dataset_accessor";
 import { checkLandmarksForThinPlateSpline } from "oxalis/model/helpers/transformation_helpers";
 import type { OxalisState } from "oxalis/store";
 import React, { useState } from "react";
