+++ frontend/javascripts/dashboard/dataset/dataset_settings_viewconfig_tab.tsx
@@ -223,8 +223,8 @@ export default function DatasetSettingsViewConfigTab(props: {
         <Col span={6}>
           <FormItemWithInfo
             name={["defaultConfiguration", "rotation"]}
-            label="Rotation - Arbitrary View Modes"
-            info="The default rotation that will be used in oblique and flight view mode."
+            label="Rotation"
+            info="The default rotation that will be used in oblique and arbitrary view mode."
           >
             <Vector3Input />
           </FormItemWithInfo>
