+++ frontend/javascripts/oxalis/geometries/materials/plane_material_factory.ts
@@ -24,13 +24,11 @@ import {
   getMagInfoByLayer,
   getMappingInfoForSupportedLayer,
   getSegmentationLayerWithMappingSupport,
-  getVisibleSegmentationLayer,
-} from "oxalis/model/accessors/dataset_accessor";
-import {
   getTransformsForLayer,
   getTransformsPerLayer,
+  getVisibleSegmentationLayer,
   invertAndTranspose,
-} from "oxalis/model/accessors/dataset_layer_transformation_accessor";
+} from "oxalis/model/accessors/dataset_accessor";
 import {
   getActiveMagIndicesForLayers,
   getUnrenderableLayerInfosForCurrentZoom,
@@ -247,7 +245,8 @@ class PlaneMaterialFactory {
     this.uniforms.activeMagIndices = {
       value: Object.values(activeMagIndices),
     };
-    const { nativelyRenderedLayerName } = Store.getState().datasetConfiguration;
+    const nativelyRenderedLayerName =
+      Store.getState().datasetConfiguration.nativelyRenderedLayerName;
     const dataset = Store.getState().dataset;
     for (const dataLayer of Model.getAllLayers()) {
       const layerName = sanitizeName(dataLayer.name);
