+++ frontend/javascripts/oxalis/model/accessors/dataset_accessor.ts
@@ -1,11 +1,18 @@
 import ErrorHandling from "libs/error_handling";
 import { formatExtentInUnitWithLength, formatNumberToLength } from "libs/format_utils";
-import { V3 } from "libs/mjs";
+import { M4x4, type Matrix4x4, V3 } from "libs/mjs";
+import MultiKeyMap from "libs/multi_key_map";
 import { aggregateBoundingBox, maxValue } from "libs/utils";
 import _ from "lodash";
 import memoizeOne from "memoize-one";
 import messages from "messages";
-import { LongUnitToShortUnitMap, type Vector3, type ViewMode } from "oxalis/constants";
+import {
+  IdentityTransform,
+  LongUnitToShortUnitMap,
+  type Vector3,
+  type Vector4,
+  type ViewMode,
+} from "oxalis/constants";
 import constants, { ViewModeValues, Vector3Indicies, MappingStatusEnum } from "oxalis/constants";
 import type {
   ActiveMappingInfo,
@@ -22,12 +29,21 @@ import type {
   APIDatasetCompact,
   APIMaybeUnimportedDataset,
   APISegmentationLayer,
+  APISkeletonLayer,
   AdditionalAxis,
   ElementClass,
 } from "types/api_flow_types";
 import type { DataLayer } from "types/schemas/datasource.types";
 import BoundingBox from "../bucket_data_handling/bounding_box";
 import { MagInfo, convertToDenseMag } from "../helpers/mag_info";
+import {
+  type Transform,
+  chainTransforms,
+  createAffineTransformFromMatrix,
+  createThinPlateSplineTransform,
+  invertTransform,
+  transformPointUnscaled,
+} from "../helpers/transformation_helpers";
 
 function _getMagInfo(magnifications: Array<Vector3>): MagInfo {
   return new MagInfo(magnifications);
@@ -698,6 +714,188 @@ export function getMappingInfoForSupportedLayer(state: OxalisState): ActiveMappi
     layer ? layer.name : null,
   );
 }
+
+// Returns the transforms (if they exist) for a layer as
+// they are defined in the dataset properties.
+function _getOriginalTransformsForLayerOrNull(
+  dataset: APIDataset,
+  layer: APIDataLayer,
+): Transform | null {
+  const coordinateTransformations = layer.coordinateTransformations;
+  if (!coordinateTransformations || coordinateTransformations.length === 0) {
+    return null;
+  }
+  if (coordinateTransformations.length > 1) {
+    console.error(
+      "Data layer has defined multiple coordinate transforms. This is currently not supported and ignored",
+    );
+    return null;
+  }
+  const transformation = coordinateTransformations[0];
+  const { type } = transformation;
+
+  if (type === "affine") {
+    const nestedMatrix = transformation.matrix;
+    return createAffineTransformFromMatrix(nestedMatrix);
+  } else if (type === "thin_plate_spline") {
+    const { source, target } = transformation.correspondences;
+
+    return createThinPlateSplineTransform(source, target, dataset.dataSource.scale.factor);
+  }
+
+  console.error(
+    "Data layer has defined a coordinate transform that is not affine or thin_plate_spline. This is currently not supported and ignored",
+  );
+  return null;
+}
+
+function _getTransformsForLayerOrNull(
+  dataset: APIDataset,
+  layer: APIDataLayer | APISkeletonLayer,
+  nativelyRenderedLayerName: string | null,
+): Transform | null {
+  if (layer.category === "skeleton") {
+    return getTransformsForSkeletonLayerOrNull(dataset, nativelyRenderedLayerName);
+  }
+  const layerTransforms = _getOriginalTransformsForLayerOrNull(dataset, layer);
+
+  if (nativelyRenderedLayerName == null) {
+    // No layer is requested to be rendered natively. Just use the transforms
+    // as they are in the dataset.
+    return layerTransforms;
+  }
+
+  if (nativelyRenderedLayerName === layer.name) {
+    // This layer should be rendered without any transforms.
+    return null;
+  }
+
+  // Apply the inverse of the layer that should be rendered natively
+  // to the current layers transforms
+  const nativeLayer = getLayerByName(dataset, nativelyRenderedLayerName, true);
+
+  const transformsOfNativeLayer = _getOriginalTransformsForLayerOrNull(dataset, nativeLayer);
+
+  if (transformsOfNativeLayer == null) {
+    // The inverse of no transforms, are no transforms. Leave the layer
+    // transforms untouched.
+    return layerTransforms;
+  }
+
+  const inverseNativeTransforms = invertTransform(transformsOfNativeLayer);
+  return chainTransforms(layerTransforms, inverseNativeTransforms);
+}
+
+function memoizeWithThreeKeys<A, B, C, T>(fn: (a: A, b: B, c: C) => T) {
+  const map = new MultiKeyMap<A | B | C, T, [A, B, C]>();
+  return (a: A, b: B, c: C): T => {
+    let res = map.get([a, b, c]);
+    if (res === undefined) {
+      res = fn(a, b, c);
+      map.set([a, b, c], res);
+    }
+    return res;
+  };
+}
+
+export const getTransformsForLayerOrNull = memoizeWithThreeKeys(_getTransformsForLayerOrNull);
+export function getTransformsForLayer(
+  dataset: APIDataset,
+  layer: APIDataLayer | APISkeletonLayer,
+  nativelyRenderedLayerName: string | null,
+): Transform {
+  return (
+    getTransformsForLayerOrNull(dataset, layer, nativelyRenderedLayerName || null) ||
+    IdentityTransform
+  );
+}
+
+function _getTransformsForSkeletonLayerOrNull(
+  dataset: APIDataset,
+  nativelyRenderedLayerName: string | null,
+): Transform | null {
+  if (nativelyRenderedLayerName == null) {
+    // No layer is requested to be rendered natively. We can use
+    // each layer's transforms as is. The skeleton layer doesn't have
+    // a transforms property currently, which is why we return null.
+    return null;
+  }
+
+  // Compute the inverse of the layer that should be rendered natively
+  const nativeLayer = getLayerByName(dataset, nativelyRenderedLayerName, true);
+  const transformsOfNativeLayer = _getOriginalTransformsForLayerOrNull(dataset, nativeLayer);
+
+  if (transformsOfNativeLayer == null) {
+    // The inverse of no transforms, are no transforms
+    return null;
+  }
+
+  return invertTransform(transformsOfNativeLayer);
+}
+
+export const getTransformsForSkeletonLayerOrNull = memoizeOne(_getTransformsForSkeletonLayerOrNull);
+
+export function getTransformsForSkeletonLayer(
+  dataset: APIDataset,
+  nativelyRenderedLayerName: string | null,
+): Transform {
+  return (
+    getTransformsForSkeletonLayerOrNull(dataset, nativelyRenderedLayerName || null) ||
+    IdentityTransform
+  );
+}
+
+function _getTransformsPerLayer(
+  dataset: APIDataset,
+  nativelyRenderedLayerName: string | null,
+): Record<string, Transform> {
+  const transformsPerLayer: Record<string, Transform> = {};
+  const layers = dataset.dataSource.dataLayers;
+  for (const layer of layers) {
+    const transforms = getTransformsForLayer(dataset, layer, nativelyRenderedLayerName);
+    transformsPerLayer[layer.name] = transforms;
+  }
+
+  return transformsPerLayer;
+}
+
+export const getTransformsPerLayer = memoizeOne(_getTransformsPerLayer);
+
+export function getInverseSegmentationTransformer(
+  state: OxalisState,
+  segmentationLayerName: string,
+) {
+  const { dataset } = state;
+  const { nativelyRenderedLayerName } = state.datasetConfiguration;
+  const layer = getLayerByName(dataset, segmentationLayerName);
+  const segmentationTransforms = getTransformsForLayer(dataset, layer, nativelyRenderedLayerName);
+  return transformPointUnscaled(invertTransform(segmentationTransforms));
+}
+
+export const hasDatasetTransforms = memoizeOne((dataset: APIDataset) => {
+  const layers = dataset.dataSource.dataLayers;
+  return layers.some((layer) => _getOriginalTransformsForLayerOrNull(dataset, layer) != null);
+});
+
+export function flatToNestedMatrix(matrix: Matrix4x4): [Vector4, Vector4, Vector4, Vector4] {
+  return [
+    matrix.slice(0, 4) as Vector4,
+    matrix.slice(4, 8) as Vector4,
+    matrix.slice(8, 12) as Vector4,
+    matrix.slice(12, 16) as Vector4,
+  ];
+}
+
+// Transposition is often needed so that the matrix has the right format
+// for matrix operations (e.g., on the GPU; but not for ThreeJS).
+// Inversion is needed when the position of an "output voxel" (e.g., during
+// rendering in the fragment shader) needs to be mapped to its original
+// data position (i.e., how it's stored without the transformation).
+// Without the inversion, the matrix maps from stored position to the position
+// where it should be rendered.
+export const invertAndTranspose = _.memoize((mat: Matrix4x4) => {
+  return M4x4.transpose(M4x4.inverse(mat));
+});
 
 export function getEffectiveIntensityRange(
   dataset: APIDataset,
