+++ frontend/javascripts/oxalis/model/accessors/flycam_accessor.ts
@@ -1,5 +1,5 @@
 import type { Matrix4x4 } from "libs/mjs";
-import { M4x4, V3 } from "libs/mjs";
+import { M4x4 } from "libs/mjs";
 import { map3, mod } from "libs/utils";
 import _ from "lodash";
 import memoizeOne from "memoize-one";
@@ -20,6 +20,8 @@ import {
   getLayerByName,
   getMagInfo,
   getMaxZoomStep,
+  getTransformsForLayer,
+  invertAndTranspose,
 } from "oxalis/model/accessors/dataset_accessor";
 import { getViewportRects } from "oxalis/model/accessors/view_mode_accessor";
 import determineBucketsForFlight from "oxalis/model/bucket_data_handling/bucket_picker_strategies/flight_bucket_picker";
@@ -33,15 +35,8 @@ import * as THREE from "three";
 import type { AdditionalCoordinate, VoxelSize } from "types/api_flow_types";
 import { baseDatasetViewConfiguration } from "types/schemas/dataset_view_configuration.schema";
 import type { SmallerOrHigherInfo } from "../helpers/mag_info";
-import {
-  type Transform,
-  chainTransforms,
-  invertTransform,
-  transformPointUnscaled,
-} from "../helpers/transformation_helpers";
 import { getMatrixScale, rotateOnAxis } from "../reducers/flycam_reducer";
 import { reuseInstanceOnEquality } from "./accessor_helpers";
-import { getTransformsForLayer, invertAndTranspose } from "./dataset_layer_transformation_accessor";
 
 export const ZOOM_STEP_INTERVAL = 1.1;
 
@@ -199,7 +194,7 @@ const perLayerFnCache: Map<string, typeof _getMaximumZoomForAllMags> = new Map()
 // Only exported for testing.
 export const _getDummyFlycamMatrix = memoizeOne((scale: Vector3) => {
   const scaleMatrix = getMatrixScale(scale);
-  return rotateOnAxis(M4x4.scale(scaleMatrix, M4x4.identity(), []), Math.PI, [0, 0, 1]);
+  return rotateOnAxis(M4x4.scale(scaleMatrix, M4x4.identity, []), Math.PI, [0, 0, 1]);
 });
 
 export function getMoveOffset(state: OxalisState, timeFactor: number) {
@@ -256,36 +251,6 @@ function getMaximumZoomForAllMagsFromStore(state: OxalisState, layerName: string
     // this already proved to be fine, though.
     dummyFlycamMatrix,
   );
-}
-
-// This function depends on functionality from this and the dataset_layer_transformation_accessor module.
-// To avoid cyclic dependencies and as the result of the function is a position and scale change,
-// this function is arguably semantically closer to this flycam module.
-export function getNewPositionAndZoomChangeFromTransformationChange(
-  activeTransformation: Transform,
-  nextTransform: Transform,
-  state: OxalisState,
-) {
-  // Calculate the difference between the current and the next transformation.
-  const currentTransformInverted = invertTransform(activeTransformation);
-  const changeInAppliedTransformation = chainTransforms(currentTransformInverted, nextTransform);
-
-  const currentPosition = getPosition(state.flycam);
-  const newPosition = transformPointUnscaled(changeInAppliedTransformation)(currentPosition);
-
-  // Also transform a reference coordinate to determine how the scaling
-  // changed. Then, adapt the zoom accordingly.
-
-  const referenceOffset: Vector3 = [10, 10, 10];
-  const secondPosition = V3.add(currentPosition, referenceOffset, [0, 0, 0]);
-  const newSecondPosition = transformPointUnscaled(changeInAppliedTransformation)(secondPosition);
-
-  const scaleChange = _.mean(
-    // Only consider XY for now to determine the zoom change (by slicing from 0 to 2)
-    V3.abs(V3.divide3(V3.sub(newPosition, newSecondPosition), referenceOffset)).slice(0, 2),
-  );
-
-  return { newPosition, scaleChange };
 }
 
 function _getUp(flycam: Flycam): Vector3 {
