+++ frontend/javascripts/oxalis/model/bucket_data_handling/layer_rendering_manager.ts
@@ -14,6 +14,8 @@ import {
   getElementClass,
   getLayerByName,
   getMagInfo,
+  getTransformsForLayer,
+  invertAndTranspose,
   isLayerVisible,
 } from "oxalis/model/accessors/dataset_accessor";
 import type { DataBucket } from "oxalis/model/bucket_data_handling/bucket";
@@ -26,10 +28,6 @@ import AsyncBucketPickerWorker from "oxalis/workers/async_bucket_picker.worker";
 import { createWorker } from "oxalis/workers/comlink_wrapper";
 import type * as THREE from "three";
 import type { AdditionalCoordinate } from "types/api_flow_types";
-import {
-  getTransformsForLayer,
-  invertAndTranspose,
-} from "../accessors/dataset_layer_transformation_accessor";
 import { getViewportRects } from "../accessors/view_mode_accessor";
 import { getSegmentsForLayer } from "../accessors/volumetracing_accessor";
 import { listenToStoreProperty } from "../helpers/listener_helpers";
