+++ frontend/javascripts/oxalis/model/helpers/transformation_helpers.ts
@@ -2,9 +2,9 @@ import { estimateAffineMatrix4x4 } from "libs/estimate_affine";
 import { M4x4 } from "libs/mjs";
 import TPS3D from "libs/thin_plate_spline";
 import type { Matrix4x4 } from "mjs";
-import type { NestedMatrix4, Vector3 } from "oxalis/constants";
+import type { Vector3, Vector4 } from "oxalis/constants";
 
-export function nestedToFlatMatrix(matrix: NestedMatrix4): Matrix4x4 {
+export function nestedToFlatMatrix(matrix: [Vector4, Vector4, Vector4, Vector4]): Matrix4x4 {
   return [...matrix[0], ...matrix[1], ...matrix[2], ...matrix[3]];
 }
 
@@ -26,7 +26,9 @@ export type Transform =
       scaledTps: TPS3D;
     };
 
-export function createAffineTransformFromMatrix(nestedMatrix: NestedMatrix4): Transform {
+export function createAffineTransformFromMatrix(
+  nestedMatrix: [Vector4, Vector4, Vector4, Vector4],
+): Transform {
   const affineMatrix = nestedToFlatMatrix(nestedMatrix);
   return { type: "affine", affineMatrix, affineMatrixInv: M4x4.inverse(affineMatrix) };
 }
