+++ frontend/javascripts/oxalis/view/context_menu.tsx
@@ -1047,7 +1047,6 @@ function getNoNodeContextMenuOptions(props: NoNodeContextMenuProps): ItemType[] 
                   : null}
               </>
             ),
-            disabled: isSkeletonLayerTransformed(state),
           },
           {
             key: "load-agglomerate-skeleton",
