+++ frontend/javascripts/types/schemas/dataset_view_configuration.schema.ts
@@ -150,9 +150,6 @@ export const datasetViewConfiguration = {
       type: "string",
     },
   },
-  nativelyRenderedLayerName: {
-    type: "string",
-  },
 };
 export default {
   $schema: "http://json-schema.org/draft-06/schema#",
